Register "Export MIDI..."; break FileMenus -> NoteTrack
